Update to 20201225
- squashed fixing DTD to main mms. Was in fact creating a problem
and then fixing it.
- squashed 'Removed non-functional MMS AP for Vodafone Portugal'
to main mms which was introducing it.
- dropped 'MMS settings for Drei' which were actually just removing
some three.co.uk apns from it. Those were introduced in upstream
MMS settings commit but not included in local MMS settings commit.
Not sure if we that delta. If we'd do, that shouldn't be in this
patch anyway.